backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0  # Changed to include WebSocket support
websockets
pydantic==2.5.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.0
google-genai
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1